+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add several overloaded methods to `MemberChecks` that assume a `null` value for `alternativeSuppressWarnings`.
+
 ### [v0.07](https://github.com/realityforge/proton/tree/v0.07) (2019-12-29) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.06...v0.07)
 
 Changes in this release: